Initialize in a more straightforward manner
1) Avoid double initialization via set_greyscale():

Since esphome initialization in main.cpp looks like this:

  App.register_component(t547_t547);
  ...
  t547_t547->set_greyscale(false);
  ...
  App.setup();

...`T547::initialize_()` used to be called twice (first from set_greyscale(),
second from setup()). This happened to call epd_init() only the second time,
because it's wrapped in `if (this->buffer_ != nullptr) {`.

This works great, but was very confusing during debugging.

2) Avoid drawing to the screen in setup().

This used to clear the screen twice during setup(): first via clean(),
then via display() of an empty buffer. It is enough to init the buffer
and the epd and wait for esphome to ask us to `update()`.